fix : 카메라 부분 유아이 수정중
