--- lib/lib/constant/colors.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-const Color BACKGROUND_COLOR=Color(0XFFF6FCFF);
-const Color BACKGROUND_SECOND_COLOR=Color(0XFFFFF9F7);
-const Color MAINSCREEN_COLOR=Color(0XFFEDFFF8);
-const Color FIRSTSCREEN_COLOR=Color(0XFF6287AA);
-const Color DEFAULT_YELLOW=Color(0XFFFDFFEF);
-const Color BUTTON_BLUE=Color(0XFF97ACB8);
-const Color BUTTON_WHITE=Color(0XFFFEFDF8);
-const Color BUTTON_PINK=Color(0XFFEBD6DC);
-const Color SHADOW_BLUE=Color(0XFF678DAD);
-const Color CHART_RED=Color(0XB3FEAEAE);
-const Color CHART_BLUE=Color(0xB3C6D9FE);
-const Color CHART_GREEN=Color(0XB3C3F7DE);
