--- lib/lib/constant/enums.dart
@@ -1,8 +1,0 @@
-enum SpaceCategoryType{
-  house,
-  office,
-  cafe,
-  etc,
-  all,
-}
-
